miercoles 14/07

+++ 1-EntradaSalida/jsEntradaSalida-08.js
@@ -11,9 +11,9 @@ function SacarResto()
 	dividendo = document.getElementById ("txtIdNumeroDividendo").value;
 	divisor = document.getElementById ("txtIdNumeroDivisor").value;
 	//parseo
-	dividendo = parseFloat(dividendo);
-	divisor = parseFloat(divisor);
+	dividendo = parseInt(dividendo);
+	divisor = parseInt(divisor);
 	//hago la operacion en alert
 
-       alert("El resto es "+(dividendo%divisor));
+       alert("El resto es " + (dividendo%divisor));
 }
